Tatt kode ut i egen komponent SøknadTittelOgIkon for å gjøre koden mer lesbar. Stylet teksten med brevkode og dato slik at det blir likt med designet.
